Game manager intrduction
- Added a game manager to handle the texture update separately from the rest.
